requirements.txt
crawl4ai
pydantic
python-dotenv
beautifulsoup4
lxml
aiohttp